add fix purchase sums, fix serializers, change decimal to integer units, add purchases test

+++ db/migrate/20220904010152_create_purchases.rb
@@ -3,7 +3,7 @@ class CreatePurchases < ActiveRecord::Migration[7.0]
     create_table :purchases do |t|
       t.references :client, null: false, foreign_key: true
       t.integer :total_products
-      t.decimal :total_amount
+      t.integer :total_amount
       t.datetime :date
 
       t.timestamps
