+++ db/migrate/20220904010427_create_purchase_items.rb
@@ -4,7 +4,7 @@ class CreatePurchaseItems < ActiveRecord::Migration[7.0]
       t.references :purchase, null: false, foreign_key: true
       t.references :product, null: false, foreign_key: true
       t.integer :quantity
-      t.decimal :total
+      t.integer :total
       t.timestamps
     end
   end
